==> analysis/ord_sil/02_data_plots.Rmd <==
# Explore data

Now that's I've cleaned up the data a bunch, I'm going to explore some of the patterns in the data. I'm interested in how occurrence changes over time, when multiple last occurrences are in the same time bin, and geological unit lithological composition.

```{r import_clean, message = FALSE, results = 'hide'}

occur_collect <- read_csv(here::here('data', 'occurrences_clean.csv'))

```


## Shelly occurrences 

The shelly groups are a mixture of phyla and classes. I want a single vector with the shelly identity of every observation.

```{r shelly}

occur_collect %<>%
  mutate(phylum_shelly = ifelse(phylum %in% shelly, phylum, NA),
         class_shelly = ifelse(class %in% shelly, class, NA),
         shelly = coalesce(phylum_shelly, class_shelly))

```

With this information I can plot occurrences over time, highlighted by their shelly type.

```{r vis_shelly}

occur_collect %>%
  group_by(shelly, bin_age) %>%
  count() %>%
  ggplot(aes(x = bin_age, y = n, 
             colour = shelly, fill = shelly)) +
  geom_col() +
  scale_fill_viridis(discrete = TRUE) +
  scale_color_viridis(discrete = TRUE) +
  scale_x_reverse() +
  labs(x = 'Time (Mya)', y = 'Count')

occur_collect %>%
  group_by(shelly, bin_age) %>%
  count() %>%
  ggplot(aes(x = bin_age, y = n, 
             colour = shelly, fill = shelly)) +
  geom_col(position = 'fill') +
  scale_fill_viridis(discrete = TRUE) +
  scale_color_viridis(discrete = TRUE) +
  scale_x_reverse() +
  labs(x = 'Time (Mya)', y = 'Proportion')

```

The oldest intervals in our dataset are mostly trilobites, while younger in the series there are more brachiopods.  There are some intervals that are absolutely dominated by individual groups.Additionally, corals become more present in younger intervals.

However, here's the catch -- the above plots are for all occurrences in the sense of abundance, and not unique taxonomic occurrences. Here's a redo of the above that focuses only on unique taxon occurrences. I'm going to filter to genus because that's such a thing in paleo.


```{r vis_shelly_unique}

occur_collect %>%
  group_by(shelly, bin_age) %>%
  filter(!duplicated(genus)) %>%
  count() %>%
  ggplot(aes(x = bin_age, y = n, 
             colour = shelly, fill = shelly)) +
  geom_col() +
  scale_fill_viridis(discrete = TRUE) +
  scale_color_viridis(discrete = TRUE) +
  scale_x_reverse() +
  labs(x = 'Time (Mya)', y = 'Count')

occur_collect %>%
  group_by(shelly, bin_age) %>%
  filter(!duplicated(genus)) %>%
  count() %>%
  ggplot(aes(x = bin_age, y = n, 
             colour = shelly, fill = shelly)) +
  geom_col(position = 'fill') +
  scale_fill_viridis(discrete = TRUE) +
  scale_color_viridis(discrete = TRUE) +
  scale_x_reverse() +
  labs(x = 'Time (Mya)', y = 'Proportion')

```

Mostly the same story as before, but some of the patterns we observed earlier are no longer as extreme. Makes sense because we're moving from abundance to incidence. 


## Last occurrences

When is the last time each genus occurs in our dataset? Group the tibble by genus and the calculate the *minimum* time bin. Minimum because time bin age *decreases* as we approach the present.

```{r lad}

lad <- 
  occur_collect %>%
  group_by(shelly, bin_age) %>%
  filter(!duplicated(genus)) %>%
  ungroup() %>%
  group_by(genus) %>%
  dplyr::summarize(lad = min(bin_age)) %>%
  group_by(lad) %>%
  count() %>%
  ungroup() %>%
  arrange(desc(lad)) %>%
  mutate(n_cum = cumsum(n))

lad %>%
  ggplot(aes(x = lad, y = n)) +
  geom_line(size = 2) +
  scale_x_reverse() +
  labs(x = 'Time (Mya)', y = 'LAD Count')

lad %>%
  ggplot(aes(x = lad, y = n_cum)) +
  geom_line(size = 2) +
  scale_x_reverse() +
  labs(x = 'Time (Mya)', y = 'LAD Cummulative')

```


And again by shelly group.

```{r lad_shelly}

lad_shelly <- 
  occur_collect %>%
  group_by(shelly, bin_age) %>%
  filter(!duplicated(genus)) %>%
  ungroup() %>%
  group_by(genus, shelly) %>%
  dplyr::summarize(lad = min(bin_age)) %>%
  group_by(shelly) %>%
  ungroup() %>%
  group_by(shelly, lad) %>%
  count() %>%
  ungroup() %>%
  group_by(shelly) %>%
  arrange(desc(lad)) %>%
  mutate(n_cum = cumsum(n)) %>%
  ungroup()

lad_shelly %>%
  ggplot(aes(x = lad, y = n, colour = shelly)) +
  geom_line(size = 2) +
  scale_colour_viridis(discrete = TRUE) +
  scale_x_reverse() +
  labs(x = 'Time (Mya)', y = 'LAD Count')

lad_shelly %>%
  ggplot(aes(x = lad, y = n_cum, colour = shelly)) +
  geom_line(size = 2) +
  scale_colour_viridis(discrete = TRUE) +
  scale_x_reverse() +
  labs(x = 'Time (Mya)', y = 'LAD Cummulative')

```

While not the largest spike, all of the shelly groups have spikes in number of LADs at the Hirnantian. How do we interptet this in light of our very biased occurrence record?


## Lithology

For each geological unit there is a short lithological description. Each lithological desciption can have multiple parts which are given a percent of total composition. Here is an example:

```{r ex_lithology}

occur_collect$lith[1]

```

In this example, the lithology of this unit is made up of three major forms with each being approximately one-third of that units composition. Lucky for me, these descriptions have a very regular form which makes extracting important text information surprisingly easy.

There are a lot of potential issues when processing lithological descriptions. For example:

- lots of words that appear only once
- similar word but different part of speech
- descriptions are identical except for small modified ("black shale siliciclastic" vs "red shale siliciclastic")

All of these problems mean I have to simplify these descriptions. 

A basic way of dividing lithologies are by siliciclsatic vs carbonate. Unfortunately, this comparison is not truly binary as mixed-sediments are possible. Additionally, I might be able to extract better information that just "siliciclastic" from our descriptions. With this in mind, I've written a function to simplify lithological descriptions to be more amenable to analysis. My goal is to divide the litholgies into "coarse siliciclastic," "fine siliciclastic," "carbonate," and "other."

```{r get_lithology}

# lithologies are now a list-column
occur_collect %<>%
  mutate(lith_simple = simplify_lithology(lith)) %>%
  unnest() %>%
  mutate(carbonate = replace_na(carbonate, replace = 0),
         coarse_siliciclastic = replace_na(coarse_siliciclastic, replace = 0),
         fine_siliciclastic = replace_na(fine_siliciclastic, replace = 0),
         other = replace_na(other, replace = 0)) %>%
  nest(carbonate, coarse_siliciclastic, fine_siliciclastic, other,
       .key = 'lith_simple')

```

I've created a column vector `lith_simple` where each element is a tibble describing the lithologycal composition of that unit in a simpler form. Most of the tricks above are just to make it so each unit has all four lithological descriptors and that there are 0s instead of NAs when that descriptor wasn't measured.



Now that that data is all in order, we can start exploring that space. Does the average lithological composition of a geologic unit change over time? If so, how?

I'm making the lithological data into a [Atchinson composition](https://rdrr.io/cran/compositions/man/acomp.html) object of the data. This object type has convenient properties for interfacing with the `compositions` package. Once this is done, there are a lot of ways to try and summarize this data. From a math perspective, all the 0s in are dataset cause a lot of problems, but hopefully we'll be able to skirt those for now.


First, I'm going to calculate mean lithology for the entire sample. Then I'm going to calculate mean lithology for each of the time bins we've defined.

```{r mean_lithology}
# divide units up by bin
# then get mean lithology
bin_lith <- 
  occur_collect %>%
  arrange(desc(bin_age)) %>%
  group_by(bin_age) %>%
  filter(!duplicated(unit_id)) %>%     # no multi-counting
  ungroup() %>%
  group_by(bin_age) %>%
  group_map(~ .x %>% 
              dplyr::select(lith_simple) %>% 
              unnest() %>%
              acomp() %>%
              mean())

# what age is each value from?
bin_ages_match <- 
  occur_collect %>%
  arrange(desc(bin_age)) %>%
  group_by(bin_age) %>%
  filter(!duplicated(unit_id)) %>%     # again, no multi-counting
  ungroup() %>%
  dplyr::select(bin_age) %>% 
  pull() %>%
  unique()

lith_time <- 
  reduce(bin_lith, rbind) %>%
  as_tibble() %>%
  mutate(bin_age = bin_ages_match) %>%
  gather(key = key, value = value, -bin_age)

lith_time %>%
  ggplot(aes(x = bin_age, y = value, fill = key)) +
  geom_col(size = 1.5) +
  scale_fill_viridis(discrete = TRUE) +
  scale_colour_viridis(discrete = TRUE) +
  scale_x_reverse() +
  labs(x = 'Time (Mya)', y = 'Percent')

```

There do not appear to be any massive shifts in mean composition from time bin to time bin. Remember, some units span multiple bins -- we've divided units by occurrences. We can also have point size vary by the diversity -- the number of unique genera in that unit at that time.

```{r pca_lithology}
lith_data <- 
  occur_collect %>%
  arrange(desc(bin_age)) %>%
  group_by(bin_age) %>%
  filter(!duplicated(unit_id)) %>%     # no multi-counting
  ungroup() %>%
  dplyr::select(lith_simple) %>%
  unnest() %>%
  acomp()

lith_pca <- princomp(lith_data)

# ggfortify doesn't support princomp.acomp output
# have to assemble plot by hand

# get age and diversity labels
lith_labels <- 
  occur_collect %>%
  arrange(desc(bin_age)) %>%
  group_by(bin_age, unit_id) %>%
  mutate(div = n_distinct(genus)) %>%
  ungroup() %>%
  group_by(bin_age) %>%
  filter(!duplicated(unit_id)) %>%     # no multi-counting
  ungroup() %>%
  dplyr::select(div, bin_age)


# get loadings to add to plot
lith_load <- 
  data.frame(lith_pca$loadings[1:4, 1:3]) %>%
  rownames_to_column() %>%
  as_tibble() %>%
  clean_names() %>%
  mutate(origin = 0)

# make the plot
as_tibble(lith_pca$scores) %>%
  clean_names() %>%
  mutate(age = lith_labels$bin_age,
         div = lith_labels$div) %>%
  ggplot() +
  geom_vline(xintercept = 0, linetype = 'dashed', alpha = 0.5) +
  geom_hline(yintercept = 0, linetype = 'dashed', alpha = 0.5) +
  geom_segment(data = lith_load,
               mapping = aes(x = origin, y = origin,
                             xend = comp_1,
                             yend = comp_2),
               arrow = arrow(length = unit(0.2, 'cm')),
               size = 1.5) +
  geom_text_repel(data = lith_load,
                  mapping = aes(x = comp_1, y = comp_2, label = rowname)) +
  geom_point(mapping = aes(x = comp_1, y = comp_2, colour = age, size = div)) +
  scale_colour_viridis() +
  labs(colour = 'Age (My)',
       size = 'Diversity',
       x = 'PC 1',
       y = 'PC 2')

```

There doesn't appear to be a lot of obvious temporal or diversity pattern for the data. 


There are some other ways of looking at our PCA results. For example, we can look at the compositions implied by the loadings of the components (and its inverse). We can also look at the relative loadings of the components for each principle component on the respective ratios, both "up and down." The up si the loading, while the down in the inverse loading.

```{r pca_more}

## compositions that represent a perturbation with the vectors represented by the loadings of each of the factors
#plot(lith_pca, type = 'loadings', scale.sdev = 1)
#
## compositions that represent a perturbation with the inverse of the vectors represented by the loadings of each of the factors
#plot(lith_pca, type = 'loadings', scale.sdev = -1)
#
## relative loadings
#plot(lith_pca, type = 'relative')

```
